requirements.txt
matplotlib>=3.7
numpy>=1.24
requests>=2.28
openai>=1.0
playwright>=1.58
python-dotenv>=1.0
lumaai>=0.2
google-api-python-client>=2.100
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.1
num2words>=0.5
